fix: address review issues and fix cluster list dropdown/tooltip clipping
- Extract bin helpers to lib/bin-utils.ts to fix cross-module import direction
- Export NEVER_EXPIRE_TTL and remove duplicate constant in formatters.ts
- Restore lastUpdateMs display in record-metadata-grid
- Add data-header attr to prevent hover highlight on header rows
- Fix mobile bin-row-grid to hide index and action columns
- Add overflow-visible to cluster list name/actions columns to fix
  dropdown menu and ACKO tooltip being clipped by td overflow-hidden
