+++ frontend/src/components/cluster-list/cluster-list-columns.tsx
@@ -32,7 +32,7 @@ export function getClusterListColumns(
       header: "Cluster Name",
       size: 280,
       enableSorting: true,
-      meta: { mobileSlot: "title" as const },
+      meta: { mobileSlot: "title" as const, cellClassName: "!overflow-visible" },
       cell: ({ row }) => {
         const { name, description, isAckoManaged } = row.original;
         return (
@@ -183,7 +183,7 @@ export function getClusterListColumns(
       header: "",
       size: 60,
       enableSorting: false,
-      meta: { mobileSlot: "actions" as const },
+      meta: { mobileSlot: "actions" as const, cellClassName: "!overflow-visible" },
       cell: ({ row }) => {
         const { id, source } = row.original;
         return (
